+++ app/controllers/api/ui/v1/authentication_controller.rb
@@ -10,7 +10,7 @@ class Api::Ui::V1::AuthenticationController < Api::Ui::BaseController
     if service.success?
       @auth_token = service.result
       @user = service.user
-      if @user.confirmed_at.nil?y
+      if @user.confirmed_at.nil?
         render json: { error: t('ui.users.alerts.unconfirmed') }, status: :unauthorized
       else
         @user.log_devise_action('login success')