+++ db/schema.rb
@@ -120,12 +120,13 @@ ActiveRecord::Schema.define(version: 2020_01_11_235900) do
   create_table "users", force: :cascade do |t|
     t.string "firstname", limit: 30
     t.string "lastname", limit: 30
-    t.string "email", limit: 60
     t.date "birtdate"
     t.string "gender", limit: 1
     t.integer "state", limit: 2, default: 10
+    t.string "phone", limit: 15
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.string "email", default: "", null: false
     t.string "encrypted_password", default: "", null: false
     t.string "reset_password_token"
     t.datetime "reset_password_sent_at"
@@ -140,7 +141,7 @@ ActiveRecord::Schema.define(version: 2020_01_11_235900) do
     t.datetime "confirmation_sent_at"
     t.string "unconfirmed_email"
     t.index ["confirmation_token"], name: "index_users_on_confirmation_token", unique: true
-    t.index ["email"], name: "index_users_on_email"
+    t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end